[#29] refactor : infrastructure 모듈 ktlint 적용
