+++ coupon-platform-infrastructure/src/main/kotlin/coupon_platform/infrastructure/account/AccountStoreImpl.kt
@@ -2,8 +2,8 @@ package coupon_platform.infrastructure.account
 
 import coupon_platform.domain.account.AccountStore
 import coupon_platform.domain.common.CommonConstants.MAX_ACCOUNTS_NUMBER
-import coupon_platform.infrastructure.cache.global_redis.handler.RedisHandlerOfBitset
-import coupon_platform.infrastructure.cache.global_redis.util.CacheConstants.ATTENDANCE_CHECK_BITOP_RESULT_KEY_TTL
+import coupon_platform.infrastructure.cache.globalredis.handler.RedisHandlerOfBitset
+import coupon_platform.infrastructure.cache.globalredis.util.CacheConstants.ATTENDANCE_CHECK_BITOP_RESULT_KEY_TTL
 import org.springframework.stereotype.Repository
 import java.time.Duration
 import java.time.LocalDateTime
@@ -27,7 +27,7 @@ class AccountStoreImpl(
         redisHandler.set(
             key = key,
             value = findBitSet,
-            ttl = Duration.ofDays(ATTENDANCE_CHECK_BITOP_RESULT_KEY_TTL)
+            ttl = Duration.ofDays(ATTENDANCE_CHECK_BITOP_RESULT_KEY_TTL),
         )
     }
