+++ coupon-platform-infrastructure/src/main/kotlin/coupon_platform/infrastructure/coupon/CouponStoreImpl.kt
@@ -10,5 +10,4 @@ class CouponStoreImpl(
 ) : CouponStore {
 
     override fun saveCoupon(coupon: Coupon): Coupon = couponJpaRepository.save(coupon)
-
 }
