--- coupon-platform-infrastructure/src/main/kotlin/coupon_platform/infrastructure/coupon_code/CouponCodeJpaRepository.kt
@@ -1,8 +1,0 @@
-package coupon_platform.infrastructure.coupon_code
-
-import coupon_platform.domain.couponcode.entity.CouponCode
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface CouponCodeJpaRepository : JpaRepository<CouponCode, Long> {
-    fun findCouponCodeByCode(code: String): CouponCode?
-}
