--- coupon-platform-infrastructure/src/main/kotlin/coupon_platform/infrastructure/coupon_code/CouponCodeStoreImpl.kt
@@ -1,13 +1,0 @@
-package coupon_platform.infrastructure.coupon_code
-
-import coupon_platform.domain.couponcode.entity.CouponCode
-import coupon_platform.domain.couponcode.repository.CouponCodeStore
-import org.springframework.stereotype.Repository
-
-@Repository
-class CouponCodeStoreImpl(
-    private val couponCodeJpaRepository: CouponCodeJpaRepository,
-) : CouponCodeStore {
-
-    override fun createCouponCode(couponCode: CouponCode): CouponCode = couponCodeJpaRepository.save(couponCode)
-}
