+++ coupon-platform-infrastructure/src/main/kotlin/coupon_platform/infrastructure/couponstats/CouponDailyDailyStatsStoreImpl.kt
@@ -1,0 +1,15 @@
+package coupon_platform.infrastructure.couponstats
+
+import coupon_platform.domain.couponstats.domain.CouponDailyStats
+import coupon_platform.domain.couponstats.repository.CouponDailyStatsStore
+import org.springframework.stereotype.Repository
+
+@Repository
+class CouponDailyDailyStatsStoreImpl(
+    private val couponDailyStatsJpaRepository: CouponDailyStatsJpaRepository,
+) : CouponDailyStatsStore {
+
+    override fun saveDailyStats(couponDailyStats: List<CouponDailyStats>) {
+        couponDailyStatsJpaRepository.saveAll(couponDailyStats)
+    }
+}
